drag and drop basic layout and localstorage
